// packages/node-webcodecs/lib/types.ts
export type BufferSource = ArrayBuffer | ArrayBufferView;
// ...
export type AudioSampleFormat =
  | 'u8'
  | 's16'
  | 's32'
  | 'f32'
  | 'u8-planar'
  | 's16-planar'
  | 's32-planar'
  | 'f32-planar';
// ...
export interface AudioDataInit {
  format: AudioSampleFormat;
  sampleRate: number; // float
  numberOfFrames: number; // unsigned long
  numberOfChannels: number; // unsigned long
  timestamp: number; // long long, microseconds
  data: BufferSource;
  transfer?: ArrayBuffer[];
}

/**
 * WebIDL:
 * dictionary AudioDataCopyToOptions { ... };
 */
export interface AudioDataCopyToOptions {
  planeIndex: number; // required unsigned long
  frameOffset?: number; // unsigned long, default 0
  frameCount?: number; // unsigned long
  format?: AudioSampleFormat;
}